+++ src/main/java/com/mju/mentoring/member/exception/exceptions/CookieException.java
@@ -1,4 +1,4 @@
-package com.mju.mentoring.member.exception;
+package com.mju.mentoring.member.exception.exceptions;
 
 public class CookieException extends RuntimeException {
