+++ src/main/java/com/mju/mentoring/member/service/helper/MemberServiceHelper.java
@@ -2,7 +2,7 @@ package com.mju.mentoring.member.service.helper;
 
 import com.mju.mentoring.member.domain.Member;
 import com.mju.mentoring.member.domain.MemberRepository;
-import com.mju.mentoring.member.exception.MemberNotFoundException;
+import com.mju.mentoring.member.exception.exceptions.MemberNotFoundException;
 import com.mju.mentoring.member.service.dto.AuthRequest;
 
 public class MemberServiceHelper {
